2. main 패키지 구조 전체 변경
