+++ src/main/java/com/nexters/phochak/auth/adapter/in/web/AuthController.java
@@ -1,7 +1,9 @@
-package com.nexters.phochak.auth.presentation;
+package com.nexters.phochak.auth.adapter.in.web;
 
-import com.nexters.phochak.auth.application.AuthService;
-import com.nexters.phochak.auth.application.JwtTokenService;
+import com.nexters.phochak.auth.application.port.in.AuthProcessUseCase;
+import com.nexters.phochak.auth.application.port.in.JwtResponseDto;
+import com.nexters.phochak.auth.application.port.in.JwtTokenUseCase;
+import com.nexters.phochak.auth.application.port.in.LoginRequestDto;
 import com.nexters.phochak.post.CommonResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -16,12 +18,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/v2/auth")
 @RestController
 public class AuthController {
-    private final AuthService authService;
-    private final JwtTokenService jwtTokenService;
+    private final AuthProcessUseCase authService;
+    private final JwtTokenUseCase jwtTokenUseCase;
 
     @GetMapping("/login/{provider}")
     public CommonResponse<JwtResponseDto> loginV2(@PathVariable String provider, @Valid LoginRequestDto requestDto) {
         Long loginUserId = authService.login(provider, requestDto);
-        return new CommonResponse<>(jwtTokenService.issueToken(loginUserId));
+        return new CommonResponse<>(jwtTokenUseCase.issueToken(loginUserId));
     }
 }
