+++ src/main/java/com/nexters/phochak/auth/application/KakaoOAuthService.java
@@ -1,7 +1,8 @@
 package com.nexters.phochak.auth.application;
 
-import com.nexters.phochak.auth.OAuthUserInformation;
-import com.nexters.phochak.auth.presentation.KakaoInformationFeignClient;
+import com.nexters.phochak.auth.adapter.out.web.KakaoInformationFeignClient;
+import com.nexters.phochak.auth.application.port.in.OAuthUseCase;
+import com.nexters.phochak.auth.application.port.in.OAuthUserInformation;
 import com.nexters.phochak.user.domain.OAuthProviderEnum;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class KakaoOAuthServiceImpl implements OAuthService {
+public class KakaoOAuthService implements OAuthUseCase {
     private static final OAuthProviderEnum OAUTH_PROVIDER = OAuthProviderEnum.KAKAO;
     private static final String TOKEN_PREFIX = "Bearer ";
     private final KakaoInformationFeignClient kakaoInformationFeignClient;
