+++ src/main/java/com/nexters/phochak/auth/application/port/in/AppleUserInformation.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.auth;
+package com.nexters.phochak.auth.application.port.in;
 
 import com.nexters.phochak.user.domain.OAuthProviderEnum;
 import lombok.AllArgsConstructor;
