+++ src/main/java/com/nexters/phochak/auth/application/port/in/LogoutRequestDto.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.auth.presentation;
+package com.nexters.phochak.auth.application.port.in;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ReissueTokenRequestDto {
-    private String accessToken;
+public class LogoutRequestDto {
     private String refreshToken;
-
 }
