+++ src/main/java/com/nexters/phochak/auth/application/port/in/OAuthUseCase.java
@@ -1,9 +1,8 @@
-package com.nexters.phochak.auth.application;
+package com.nexters.phochak.auth.application.port.in;
 
-import com.nexters.phochak.auth.OAuthUserInformation;
 import com.nexters.phochak.user.domain.OAuthProviderEnum;
 
-public interface OAuthService {
+public interface OAuthUseCase {
     OAuthProviderEnum getOAuthProvider();
 
     OAuthUserInformation requestUserInformation(String token);
