+++ src/main/java/com/nexters/phochak/auth/application/port/in/ReissueTokenRequestDto.java
@@ -1,4 +1,4 @@
-package com.nexters.phochak.auth.presentation;
+package com.nexters.phochak.auth.application.port.in;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class LogoutRequestDto {
+public class ReissueTokenRequestDto {
+    private String accessToken;
     private String refreshToken;
+
 }
