+++ src/main/java/com/nexters/phochak/ignore/presentation/IgnoreController.java
@@ -1,7 +1,7 @@
 package com.nexters.phochak.ignore.presentation;
 
-import com.nexters.phochak.auth.UserContext;
-import com.nexters.phochak.auth.annotation.Auth;
+import com.nexters.phochak.auth.interceptor.Auth;
+import com.nexters.phochak.auth.interceptor.UserContext;
 import com.nexters.phochak.ignore.IgnoredUserResponseDto;
 import com.nexters.phochak.post.CommonResponse;
 import com.nexters.phochak.user.application.UserService;
