+++ src/test/java/com/nexters/phochak/auth/application/AuthServiceTest.java
@@ -1,6 +1,6 @@
 package com.nexters.phochak.auth.application;
 
-import com.nexters.phochak.auth.KakaoUserInformation;
+import com.nexters.phochak.auth.application.port.in.KakaoUserInformation;
 import com.nexters.phochak.user.domain.User;
 import com.nexters.phochak.user.domain.UserFixture;
 import com.nexters.phochak.user.domain.UserRepository;
@@ -14,7 +14,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static com.nexters.phochak.auth.KakaoUserInformation.KakaoOAuthProperties;
+import static com.nexters.phochak.auth.application.port.in.KakaoUserInformation.KakaoOAuthProperties;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
@@ -26,7 +26,7 @@ class AuthServiceTest {
     @Mock
     UserRepository userRepository;
     @InjectMocks
-    AuthServiceImpl authService;
+    AuthProcessService authService;
     KakaoOAuthProperties kakaoOAuthProperties;
     KakaoUserInformation userInformation;
