+++ src/test/java/com/nexters/phochak/deprecated/integration/AuthIntegrationTest.java
@@ -1,11 +1,11 @@
 package com.nexters.phochak.deprecated.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nexters.phochak.auth.application.JwtTokenService;
-import com.nexters.phochak.auth.domain.RefreshTokenRepository;
-import com.nexters.phochak.auth.presentation.LogoutRequestDto;
-import com.nexters.phochak.auth.presentation.ReissueTokenRequestDto;
-import com.nexters.phochak.auth.presentation.WithdrawRequestDto;
+import com.nexters.phochak.auth.adapter.out.persistence.RefreshTokenRepository;
+import com.nexters.phochak.auth.application.port.in.JwtTokenUseCase;
+import com.nexters.phochak.auth.application.port.in.LogoutRequestDto;
+import com.nexters.phochak.auth.application.port.in.ReissueTokenRequestDto;
+import com.nexters.phochak.auth.application.port.in.WithdrawRequestDto;
 import com.nexters.phochak.common.docs.RestDocs;
 import com.nexters.phochak.common.exception.CustomExceptionHandler;
 import com.nexters.phochak.hashtag.domain.Hashtag;
@@ -40,7 +40,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static com.nexters.phochak.auth.aspect.AuthAspect.AUTHORIZATION_HEADER;
+import static com.nexters.phochak.auth.interceptor.AuthAspect.AUTHORIZATION_HEADER;
 import static com.nexters.phochak.common.exception.ResCode.*;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
@@ -64,7 +64,8 @@ public class AuthIntegrationTest extends RestDocs {
     @Autowired PostRepository postRepository;
     @Autowired ShortsRepository shortsRepository;
     @Autowired HashtagRepository hashtagRepository;
-    @Autowired JwtTokenService jwtTokenService;
+    @Autowired
+    JwtTokenUseCase jwtTokenUseCase;
     @Autowired ObjectMapper objectMapper;
     @Autowired EntityManager em;
     @Autowired RefreshTokenRepository refreshTokenRepository;
@@ -87,16 +88,16 @@ public class AuthIntegrationTest extends RestDocs {
         globalUserId = user.getId();
     }
 
-    private static String createTokenStringForResponse(JwtTokenService.TokenVo accessToken) {
-        return JwtTokenService.TokenVo.TOKEN_TYPE + " " + accessToken.getTokenString();
+    private static String createTokenStringForResponse(JwtTokenUseCase.TokenVo accessToken) {
+        return JwtTokenUseCase.TokenVo.TOKEN_TYPE + " " + accessToken.getTokenString();
     }
 
     @Test
     @DisplayName("토큰 재발급 API - 재발급 성공")
     void reissueToken() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT = jwtTokenService.generateToken(globalUserId, -1000L);
-        JwtTokenService.TokenVo currentRT = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT = jwtTokenUseCase.generateToken(globalUserId, -1000L);
+        JwtTokenUseCase.TokenVo currentRT = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT.getTokenString(), currentAT.getTokenString());
 
@@ -132,8 +133,8 @@ public class AuthIntegrationTest extends RestDocs {
     @DisplayName("토큰 재발급 API - Refresh Token이 만료된 경우, Expired Token 예외가 발생한다")
     void reissueToken_RTExpired_InvalidToken() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT = jwtTokenService.generateToken(globalUserId, -2000L);
-        JwtTokenService.TokenVo currentRT = jwtTokenService.generateToken(globalUserId, -1000L);
+        JwtTokenUseCase.TokenVo currentAT = jwtTokenUseCase.generateToken(globalUserId, -2000L);
+        JwtTokenUseCase.TokenVo currentRT = jwtTokenUseCase.generateToken(globalUserId, -1000L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT.getTokenString(), currentAT.getTokenString());
 
@@ -153,8 +154,8 @@ public class AuthIntegrationTest extends RestDocs {
     @DisplayName("토큰 재발급 API - Access Token이 아직 만료되지 않은 경우, 탈취로 간주하고 Invalid Token 예외가 발생한다")
     void reissueToken_ATNotExpired_InvalidToken() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT = jwtTokenService.generateToken(globalUserId, 1000000000L);
-        JwtTokenService.TokenVo currentRT = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT = jwtTokenUseCase.generateToken(globalUserId, 1000000000L);
+        JwtTokenUseCase.TokenVo currentRT = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT.getTokenString(), currentAT.getTokenString());
 
@@ -174,13 +175,13 @@ public class AuthIntegrationTest extends RestDocs {
     @DisplayName("토큰 재발급 API - AT와 RT가 매치되지 않는 경우, 탈취로 간주하고 Invalid Token 예외가 발생한다")
     void reissueToken_ATRTNotMatched_InvalidToken() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT1 = jwtTokenService.generateToken(globalUserId, 1000000000L);
-        JwtTokenService.TokenVo currentRT1 = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT1 = jwtTokenUseCase.generateToken(globalUserId, 1000000000L);
+        JwtTokenUseCase.TokenVo currentRT1 = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT1.getTokenString(), currentAT1.getTokenString());
 
-        JwtTokenService.TokenVo currentAT2 = jwtTokenService.generateToken(globalUserId, 1000000000L);
-        JwtTokenService.TokenVo currentRT2 = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT2 = jwtTokenUseCase.generateToken(globalUserId, 1000000000L);
+        JwtTokenUseCase.TokenVo currentRT2 = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT2.getTokenString(), currentAT2.getTokenString());
 
@@ -201,8 +202,8 @@ public class AuthIntegrationTest extends RestDocs {
     @DisplayName("로그아웃 API - 로그아웃 성공")
     void logout_success() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT1 = jwtTokenService.generateToken(globalUserId, 1000000000L);
-        JwtTokenService.TokenVo currentRT1 = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT1 = jwtTokenUseCase.generateToken(globalUserId, 1000000000L);
+        JwtTokenUseCase.TokenVo currentRT1 = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         refreshTokenRepository.saveWithAccessToken(currentRT1.getTokenString(), currentAT1.getTokenString());
 
@@ -210,7 +211,7 @@ public class AuthIntegrationTest extends RestDocs {
 
         //when, then
         mockMvc.perform(post("/v1/user/logout")
-                        .header(AUTHORIZATION_HEADER, JwtTokenService.TokenVo.TOKEN_TYPE + " " + currentAT1.getTokenString())
+                        .header(AUTHORIZATION_HEADER, JwtTokenUseCase.TokenVo.TOKEN_TYPE + " " + currentAT1.getTokenString())
                         .content(objectMapper.writeValueAsString(body))
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
@@ -232,8 +233,8 @@ public class AuthIntegrationTest extends RestDocs {
     @DisplayName("회원탈퇴 API - 회원탈퇴 성공")
     void withdraw_success() throws Exception {
         //given
-        JwtTokenService.TokenVo currentAT = jwtTokenService.generateToken(globalUserId, 1000000000L);
-        JwtTokenService.TokenVo currentRT = jwtTokenService.generateToken(globalUserId, 9999999999L);
+        JwtTokenUseCase.TokenVo currentAT = jwtTokenUseCase.generateToken(globalUserId, 1000000000L);
+        JwtTokenUseCase.TokenVo currentRT = jwtTokenUseCase.generateToken(globalUserId, 9999999999L);
 
         User user = userRepository.findById(globalUserId).get();
 
@@ -262,7 +263,7 @@ public class AuthIntegrationTest extends RestDocs {
 
         //when, then
         mockMvc.perform(post("/v1/user/withdraw")
-                        .header(AUTHORIZATION_HEADER, JwtTokenService.TokenVo.TOKEN_TYPE + " " + currentAT.getTokenString())
+                        .header(AUTHORIZATION_HEADER, JwtTokenUseCase.TokenVo.TOKEN_TYPE + " " + currentAT.getTokenString())
                         .content(objectMapper.writeValueAsString(body))
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
